please read description
I added the field "updatedAt" to Post and modified "createdAt" to be LocalDateTime. Fixed some bugs regarding add and update methods.
Now we got authorization on all endpoints except the "/auth" ones, because I wanted to test if everything works as intended, but for testing you can uncomment line 111 in /security/SecurityConfig and comment line 112.